requirements.txt
absl-py==1.4.0
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
Authlib==1.2.1
azure-ai-textanalytics==5.2.1
azure-common==1.1.28
azure-core==1.26.3
backcall==0.2.0
beautifulsoup4==4.11.2
behave==1.2.6
bleach==6.1.0
boto3==1.26.82
botocore==1.29.82
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
decorator==5.1.1
distlib==0.3.4
Django==3.2.5
django-cors-headers==3.14.0
django-jsonfield==1.4.1
django-keycloak @ git+https://github.com/Peter-Slump/django-keycloak.git@5211fba26ae4216b4344309eb4a67da9745ada05
djangorestframework==3.14.0
ecdsa==0.18.0
executing==1.2.0
filelock==3.6.0
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
google-api-core==2.11.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-cloud-language==2.9.0
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
h5py==3.8.0
idna==3.3
ipython==8.11.0
isodate==0.6.1
jax==0.4.11
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
josepy==1.13.0
kaggle==1.5.16
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mozilla-django-oidc==3.0.0
mpmath==1.3.0
msrest==0.7.1
networkx==3.0
numpy==1.23.1
oauthlib==3.2.2
opencv-python==4.7.0.72
opensearch-py==2.2.0
opt-einsum==3.3.0
packaging==23.0
pandas==1.4.3
parse==1.19.1
parse-type==0.6.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.2
plotly==5.13.1
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.4
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-keycloak-client==0.2.3
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-aws4auth==1.2.3
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.2
stack-data==0.6.2
sympy==1.11.1
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.0
traitlets==5.9.0
typing_extensions==4.5.0
ultralytics==8.0.124
urllib3==1.26.11
virtualenv==20.14.1
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.14.1
